+++ rpc/server.go
@@ -7,10 +7,11 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 	"github.com/pokt-network/pocket/shared/modules"
+	"github.com/pokt-network/pocket/shared/modules/base_modules"
 )
 
 type rpcServer struct {
-	modules.BaseIntegratableModule
+	base_modules.IntegratableModule
 
 	logger modules.Logger
 }