+++ logger/docs/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.7] - 2023-02-10
 
-- Module embeds `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
+- Module embeds `base_modules.IntegratableModule` and `base_modules.InterruptableModule` for DRYness
 
 ## [0.0.0.6] - 2023-02-09