+++ p2p/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.26] - 2023-02-10
 
-- Modules embed `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
+- Modules embed `base_modules.IntegratableModule` and `base_modules.InterruptableModule` for DRYness
 - Deprecated `debugAddressBookProvider`
 - Added `rpcAddressBookProvider` to source the address book from the RPC server
 - Leveraging `bus` for dependency injection of the `addressBookProvider` and `currentHeightProvider`