+++ rpc/doc/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.0.11] - 2023-02-09
 
 - Updated RPC to expose the node's address book via GET /v1/p2p/address_book
-- Updated modules to embed `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
+- Updated modules to embed `base_modules.IntegratableModule` and `base_modules.InterruptableModule` for DRYness
 
 ## [0.0.0.10] - 2023-02-07