+++ telemetry/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.8] - 2023-02-09
 
-- Modules embed `modules.BaseIntegratableModule` and `modules.BaseInterruptableModule` for DRYness
+- Modules embed `base_modules.IntegratableModule` and `base_modules.InterruptableModule` for DRYness
 
 ## [0.0.0.7] - 2023-02-07